Add some updates from the code review and discussion
