+++ Math/src/main/java/org/tribuo/math/neighbour/NeighboursQuery.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.tribuo.math.neighbor;
+package org.tribuo.math.neighbour;
 
 import com.oracle.labs.mlrg.olcut.util.Pair;
 import org.tribuo.math.la.SGDVector;
@@ -22,15 +22,15 @@ import org.tribuo.math.la.SGDVector;
 import java.util.List;
 
 /**
- * An interface for nearest neighbor queries.
+ * An interface for nearest neighbour queries.
  */
-public interface NeighborsQuery {
+public interface NeighboursQuery {
 
     /**
      * Queries a set of {@link SGDVector}s to determine the k points nearest to the provided point.
      * @param point The point to determine the nearest k points for.
-     * @param k The number of neighboring points to identify.
-     * @return A list of k {@link Pair}s, where a pair contains the index of the neighboring point in the original
+     * @param k The number of neighbouring points to identify.
+     * @return A list of k {@link Pair}s, where a pair contains the index of the neighbouring point in the original
      *         data and the distance between this point and the provided point.
      */
     List<Pair<Integer, Double>> query(SGDVector point, int k);
@@ -38,19 +38,19 @@ public interface NeighborsQuery {
     /**
      * Queries a set of {@link SGDVector}s to determine the k points nearest to the provided points.
      * @param points An array of points to determine the nearest k points for.
-     * @param k The number of neighboring points to identify.
-     * @return An array containing lists of k {@link Pair}s. There is array entry for each provided point which is a
-     *         list of k pairs. Each pair contains the index of the neighboring point in the original data and the
+     * @param k The number of neighbouring points to identify.
+     * @return An list containing lists of k {@link Pair}s. There is list entry for each provided point which is a
+     *         list of k pairs. Each pair contains the index of the neighbouring point in the original data and the
      *         distance between this point and the provided point.
      */
-    List<Pair<Integer, Double>>[] query(SGDVector[] points, int k);
+    List<List<Pair<Integer, Double>>> query(SGDVector[] points, int k);
 
     /**
      * Queries a set of {@link SGDVector}s to determine the k points nearest to every point in the set.
-     * @param k The number of neighboring points to identify.
-     * @return An array containing lists of k {@link Pair}s. There is array entry for each provided point which is a
-     *         list of k pairs. Each pair contains the index of the neighboring point in the original data and the
+     * @param k The number of neighbouring points to identify.
+     * @return A list containing lists of k {@link Pair}s. There is list entry for each provided point which is a
+     *         list of k pairs. Each pair contains the index of the neighbouring point in the original data and the
      *         distance between this point and the provided point.
      */
-    List<Pair<Integer, Double>>[] queryAll(int k);
+    List<List<Pair<Integer, Double>>> queryAll(int k);
 }
